requirements.txt
filterpy==1.4.5
numpy==1.23.5
pandas==1.5.3
requests==2.28.2
gpxpy==1.5.0
fitparse==1.2.0
geopy==2.3.0
dubins~=1.0.1
setuptools~=68.2.0
folium~=0.18.0
scipy~=1.10.1
cachetools~=5.5.0
PyYAML~=6.0.2
plotly~=5.24.1